requirements.txt
Faker==13.3.3
faker-commerce==1.0.3
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
